backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
opencv-python-headless==4.9.0.80
rembg==2.0.57
mediapipe==0.10.14
reportlab==4.2.0
numpy==1.26.4
aiofiles==23.2.1
python-dotenv==1.0.1
